+++ lib/twilio-ruby/rest/verify/v2/service/verification.rb
@@ -44,9 +44,11 @@ module Twilio
             #   Voice documentation of
             #   {sendDigits}[https://www.twilio.com/docs/voice/twiml/number#attributes-sendDigits].
             # @param [String] locale Locale will automatically resolve based on phone number
-            #   country code for SMS, WhatsApp and call channel verifications. This parameter
-            #   will override the automatic locale. {See supported languages and more
-            #   information here}[https://www.twilio.com/docs/verify/supported-languages].
+            #   country code for SMS, WhatsApp, and call channel verifications. It will fallback
+            #   to English or the template’s default translation if the selected translation is
+            #   not available. This parameter will override the automatic locale resolution.
+            #   {See supported languages and more information
+            #   here}[https://www.twilio.com/docs/verify/supported-languages].
             # @param [String] custom_code A pre-generated code to use for verification. The
             #   code can be between 4 and 10 characters, inclusive.
             # @param [String] amount The amount of the associated PSD2 compliant transaction.